Correction majeure : Xbee division par 0     alléger l'inerrupt InputCapture

+++ Sources/Modules/mMotor.c
@@ -16,7 +16,6 @@
 #define FTM2_OVERFLOW_FREQUENCY 				500		
 #define N_OF							1000
 
-
 #define DT_MIN 							5000
 
 /**
@@ -166,7 +165,7 @@ void FTM2_IRQHandler()
 	    mMotor1_CaptFlancPos = 0;
 	    mMotor1_CaptFlancNeg = 0;
 	    mMotor1.aCapt = 0;
-	    mMotor1.aFreq = 0.0;
+	    mMotor1.aFreq = (float)0.0;
 	    //On vide le tableau
 	    mMotor1.aNumEchantillonsMot = 0;
 	    }
@@ -182,7 +181,7 @@ void FTM2_IRQHandler()
 	    mMotor2_CaptFlancPos = 0;
 	    mMotor2_CaptFlancNeg = 0;
 	    mMotor2.aCapt = 0;
-	    mMotor2.aFreq = 0.0;
+	    mMotor2.aFreq = (float)0.0;
 	    //On vide le tableau
 	    mMotor2.aNumEchantillonsMot = 0;
 	    }
